fix(review): propagate reqwest client-build errors instead of panicking in constructors
Closes #953

Eight production constructors previously called `.expect()` on
`reqwest::Client::builder().build()`, panicking on TLS init failure at
daemon startup. This commit hardens them to return `Result` and propagates
the error to callers.

Changed constructors (now return Result):
- `HttpSearchClient::new` / `from_config`  → `Result<Self, SearchClientError>`
- `HttpAnalyzeClient::new` / `from_config` → `Result<Self, AnalyzeClientError>`
- `SubprocessAnalyzeClient::new` / `from_config` → `Result<Self, AnalyzeClientError>`
- `GithubClient::new` / `with_timeout`     → `Result<Self, GithubError>`
- `ReqwestConfluenceTransport::new`        → `Result<Self, ContextSourceError>`
- `ReqwestJiraTransport::new`              → `Result<Self, ContextSourceError>`
- `ReqwestIssueSearch::new`               → `Result<Self, ContextSourceError>`

Caller update strategy:
- Binary / command callers (run, serve, compare, mcp/mod) propagate with
  `.map_err(|e| anyhow::anyhow!(...))? ` into their existing `anyhow::Result`.
- Pipeline callers (diff.rs, post.rs, runner.rs) propagate with `?` into
  their existing `GithubError` return types.
- MCP tools.rs maps to `ToolError::InvalidParams`.
- reporter.rs degrades gracefully: logs a warning and returns `None`.
- Context-source `from_config` methods (Confluence, Jira, GitHub Issues)
  keep their infallible `Self` return type but build permanently-disabled
  sentinel sources with no-op trait impls when TLS init fails, so the
  orchestrator's `is_enabled()` skip path handles them safely.
- `Default` impls kept as-is (panicking) with clear doc explaining that
  production code should use `new()`.

Also splits `analyze_client.rs` and `jira.rs` test sections into sibling
`*_tests.rs` files (matching the existing pattern) to keep all three
affected files under the 500-line cap; updates `.line-cap-allowlist.tsv`
accordingly (ratchet: removes the now-under-cap entry for analyze_client.rs).
