+++ .line-cap-allowlist.tsv
@@ -133,7 +133,6 @@ crates/trusty-mpm/src/telegram/mod.rs	632
 crates/trusty-mpm/src/tui/dashboard.rs	1040
 crates/trusty-mpm/src/tui/health.rs	3330
 crates/trusty-mpm/src/tui/mod.rs	763
-crates/trusty-review/src/integrations/analyze_client.rs	622
 crates/trusty-review/src/models/mod.rs	531
 crates/trusty-review/src/pipeline/runner_tests.rs	730
 crates/trusty-review/src/pipeline/verify_tests.rs	533
